Functions

Functions

SpecialFunctions.erf — Function.

erf(x)

Compute the error function of x, defined by $\frac{2}{\sqrt{\pi}} \int_0^x e^{-t^2} dt$ for arbitrary complex x.

source

SpecialFunctions.erfc — Function.

erfc(x)

Compute the complementary error function of x, defined by $1 - \operatorname{erf}(x)$.

source

SpecialFunctions.erfcx — Function.

erfcx(x)

Compute the scaled complementary error function of x, defined by $e^{x^2} \operatorname{erfc}(x)$. Note also that $\operatorname{erfcx}(-ix)$ computes the Faddeeva function $w(x)$.

source

SpecialFunctions.erfi — Function.

erfi(x)

Compute the imaginary error function of x, defined by $-i \operatorname{erf}(ix)$.

source

SpecialFunctions.dawson — Function.

dawson(x)

Compute the Dawson function (scaled imaginary error function) of x, defined by $\frac{\sqrt{\pi}}{2} e^{-x^2} \operatorname{erfi}(x)$.

source

SpecialFunctions.erfinv — Function.

erfinv(x)

Compute the inverse error function of a real x, defined by $\operatorname{erf}(\operatorname{erfinv}(x)) = x$.

source

SpecialFunctions.erfcinv — Function.

erfcinv(x)

Compute the inverse error complementary function of a real x, defined by $\operatorname{erfc}(\operatorname{erfcinv}(x)) = x$.

source

SpecialFunctions.sinint — Function.

sinint(x)

Compute the sine integral function of x, defined by $\operatorname{Si}(x) := \int_0^x\frac{\sin t}{t} dt$ for real x.

source

SpecialFunctions.cosint — Function.

cosint(x)

Compute the cosine integral function of x, defined by $\operatorname{Ci}(x) := \gamma + \log x + \int_0^x \frac{\cos t - 1}{t} dt$ for real x > 0, where $\gamma$ is the Euler-Mascheroni constant.

source

SpecialFunctions.digamma — Function.

digamma(x)

Compute the digamma function of x (the logarithmic derivative of gamma(x)).

source

SpecialFunctions.invdigamma — Function.

invdigamma(x)

Compute the inverse digamma function of x.

source

SpecialFunctions.trigamma — Function.

trigamma(x)

Compute the trigamma function of x (the logarithmic second derivative of gamma(x)).

source

SpecialFunctions.polygamma — Function.

polygamma(m, x)

Compute the polygamma function of order m of argument x (the (m+1)th derivative of the logarithm of gamma(x))

source

SpecialFunctions.airyai — Function.

airyai(x)

Airy function of the first kind $\operatorname{Ai}(x)$.

source

SpecialFunctions.airyaiprime — Function.

airyaiprime(x)

Derivative of the Airy function of the first kind $\operatorname{Ai}'(x)$.

source

SpecialFunctions.airyaix — Function.

airyaix(x)

Scaled Airy function of the first kind $\operatorname{Ai}(x) e^{\frac{2}{3} x \sqrt{x}}$. Throws DomainError for negative Real arguments.

source

SpecialFunctions.airyaiprimex — Function.

airyaiprimex(x)

Scaled derivative of the Airy function of the first kind $\operatorname{Ai}'(x) e^{\frac{2}{3} x \sqrt{x}}$. Throws DomainError for negative Real arguments.

source

SpecialFunctions.airybi — Function.

airybi(x)

Airy function of the second kind $\operatorname{Bi}(x)$.

source

SpecialFunctions.airybiprime — Function.

airybiprime(x)

Derivative of the Airy function of the second kind $\operatorname{Bi}'(x)$.

source

SpecialFunctions.airybix — Function.

airybix(x)

Scaled Airy function of the second kind $\operatorname{Bi}(x) e^{- \left| \operatorname{Re} \left( \frac{2}{3} x \sqrt{x} \right) \right|}$.

source

SpecialFunctions.airybiprimex — Function.

airybiprimex(x)

Scaled derivative of the Airy function of the second kind $\operatorname{Bi}'(x) e^{- \left| \operatorname{Re} \left( \frac{2}{3} x \sqrt{x} \right) \right|}$.

source

SpecialFunctions.besselj0 — Function.

besselj0(x)

Bessel function of the first kind of order 0, $J_0(x)$.

source

SpecialFunctions.besselj1 — Function.

besselj1(x)

Bessel function of the first kind of order 1, $J_1(x)$.

source

SpecialFunctions.besselj — Function.

besselj(nu, x)

Bessel function of the first kind of order nu, $J_\nu(x)$.

source

SpecialFunctions.besseljx — Function.

besseljx(nu, x)

Scaled Bessel function of the first kind of order nu, $J_\nu(x) e^{- | \operatorname{Im}(x) |}$.

source

SpecialFunctions.bessely0 — Function.

bessely0(x)

Bessel function of the second kind of order 0, $Y_0(x)$.

source

SpecialFunctions.bessely1 — Function.

bessely1(x)

Bessel function of the second kind of order 1, $Y_1(x)$.

source

SpecialFunctions.bessely — Function.

bessely(nu, x)

Bessel function of the second kind of order nu, $Y_\nu(x)$.

source

SpecialFunctions.besselyx — Function.

besselyx(nu, x)

Scaled Bessel function of the second kind of order nu, $Y_\nu(x) e^{- | \operatorname{Im}(x) |}$.

source

SpecialFunctions.hankelh1 — Function.

hankelh1(nu, x)

Bessel function of the third kind of order nu, $H^{(1)}_\nu(x)$.

source

SpecialFunctions.hankelh1x — Function.

hankelh1x(nu, x)

Scaled Bessel function of the third kind of order nu, $H^{(1)}_\nu(x) e^{-x i}$.

source

SpecialFunctions.hankelh2 — Function.

hankelh2(nu, x)

Bessel function of the third kind of order nu, $H^{(2)}_\nu(x)$.

source

SpecialFunctions.hankelh2x — Function.

hankelh2x(nu, x)

Scaled Bessel function of the third kind of order nu, $H^{(2)}_\nu(x) e^{x i}$.

source

SpecialFunctions.besselh — Function.

besselh(nu, [k=1,] x)

Bessel function of the third kind of order nu (the Hankel function). k is either 1 or 2, selecting hankelh1 or hankelh2, respectively. k defaults to 1 if it is omitted. (See also besselhx for an exponentially scaled variant.)

source

SpecialFunctions.besselhx — Function.

besselhx(nu, [k=1,] z)

Compute the scaled Hankel function $\exp(∓iz) H_ν^{(k)}(z)$, where $k$ is 1 or 2, $H_ν^{(k)}(z)$ is besselh(nu, k, z), and $∓$ is $-$ for $k=1$ and $+$ for $k=2$. k defaults to 1 if it is omitted.

The reason for this function is that $H_ν^{(k)}(z)$ is asymptotically proportional to $\exp(∓iz)/\sqrt{z}$ for large $|z|$, and so the besselh function is susceptible to overflow or underflow when z has a large imaginary part. The besselhx function cancels this exponential factor (analytically), so it avoids these problems.

source

SpecialFunctions.besseli — Function.

besseli(nu, x)

Modified Bessel function of the first kind of order nu, $I_\nu(x)$.

source

SpecialFunctions.besselix — Function.

besselix(nu, x)

Scaled modified Bessel function of the first kind of order nu, $I_\nu(x) e^{- | \operatorname{Re}(x) |}$.

source

SpecialFunctions.besselk — Function.

besselk(nu, x)

Modified Bessel function of the second kind of order nu, $K_\nu(x)$.

source

SpecialFunctions.besselkx — Function.

besselkx(nu, x)

Scaled modified Bessel function of the second kind of order nu, $K_\nu(x) e^x$.

source

SpecialFunctions.eta — Function.

eta(x)

Dirichlet eta function $\eta(s) = \sum^\infty_{n=1}(-1)^{n-1}/n^{s}$.

source

SpecialFunctions.zeta — Function.

zeta(s, z)

Generalized zeta function $\zeta(s, z)$, defined by the sum $\sum_{k=0}^\infty ((k+z)^2)^{-s/2}$, where any term with $k+z=0$ is excluded. For $\Re z > 0$, this definition is equivalent to the Hurwitz zeta function $\sum_{k=0}^\infty (k+z)^{-s}$. For $z=1$, it yields the Riemann zeta function $\zeta(s)$.

source

zeta(s)

Riemann zeta function $\zeta(s)$.

source